requirements.txt
autopep8==1.5.5
isort==5.7.0
pycodestyle==2.6.0
pylint==2.6.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
PyMySQL==1.0.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.5.1
Werkzeug==1.0.1
python-dotenv==0.15.0
cryptography==3.4.6